requirements.txt
aiofiles==0.5.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
cacheout==0.11.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudflare==2.8.3
ecdsa==0.15
fastapi==0.58.1
h11==0.9.0
idna==2.10
jsonlines==1.2.0
passlib==1.7.2
pyasn1==0.4.8
pycparser==2.20
pydantic==1.5.1
python-jose==3.1.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
starlette==0.13.4
urllib3==1.25.9
uvicorn==0.11.5
websockets==8.1
wincertstore==0.2
